Refactor the SimDB exporters
