+++ sparta/scripts/simdb/exporters/export_text_report.py
@@ -1,83 +1,7 @@
-import io, os, zlib, struct
+from .sim_utils import *
+from .stat_utils import *
+import io
 from .utils import FormatNumber
-
-class SimulationInfo:
-    @classmethod
-    def GetHeaderPairs(cls, db_conn):
-        cursor = db_conn.cursor()
-        cmd = "SELECT HeaderName, HeaderValue FROM SimulationInfoHeaderPairs"
-        cursor.execute(cmd)
-
-        header_pairs = []
-        for p in cursor.fetchall():
-            header_pairs.append((p[0], p[1]))
-
-        return header_pairs
-
-    @classmethod
-    def Stringize(cls, db_conn, out, line_start = "# ", line_end = "\n", show_field_names = True):
-        pairs = cls.GetHeaderPairs(db_conn)
-        for p in pairs:
-            out.write(line_start)
-            if show_field_names:
-                # out << std::left << std::setw(10) << (p.first + ":");
-                name = p[0] + ":"
-                out.write(f"{name:<10}")
-
-            # out << p.second << line_end;
-            out.write(f"{p[1]}{line_end}")
-
-class StatisticInstance:
-    def __init__(self, name, location, desc, stat_value_getter):
-        self.name = name
-        self.location = location
-        self.desc = desc
-        self.stat_value_getter = stat_value_getter
-
-    def GetName(self):
-        return self.name
-    
-    def GetLocation(self):
-        return self.location
-
-    def GetDesc(self, unused=False):
-        assert unused == False
-        return self.desc
-
-    def GetValue(self):
-        return self.stat_value_getter.GetNext()
-
-class StatValueGetter:
-    def __init__(self, stats_values):
-        self.stats_values = stats_values
-        self.index = 0
-
-    def GetNext(self):
-        if self.index >= len(self.stats_values):
-            raise IndexError("No more values in stats blob")
-        value = self.stats_values[self.index]
-        self.index += 1
-        return value
-
-def GetStatsValuesGetter(cursor, dest_file):
-    dest_file_name = os.path.basename(dest_file)
-    cmd = f"SELECT Data, IsCompressed FROM CollectionRecords WHERE Notes='{dest_file_name}'"
-    cursor.execute(cmd)
-    stats_blob, is_compressed = cursor.fetchone()
-    if is_compressed:
-        stats_blob = zlib.decompress(stats_blob)
-
-    # Turn the stats blob (byte vector) into a vector of doubles.
-    assert len(stats_blob) % (2+8) == 0, "Invalid stats blob length"
-    stats_values = []
-    for i in range(0, len(stats_blob), 2+8):
-        # The first value is the collectable ID, the second is the value.
-        # We don't need the collectable ID here.
-        val = struct.unpack("d", stats_blob[i+2:i+10])[0]
-        val = FormatNumber(val, as_string=False)
-        stats_values.append(val)
-
-    return StatValueGetter(stats_values)
 
 class TextReportExporter:
     DEFAULT_REPORT_PREFIX = "Report "
